feat(overview): remove overview page
